requirements.txt
scrapy-zyte-api
zyte-spider-templates